+++ sdk/resourcemanager/compute/armcompute/CHANGELOG.md
@@ -1,9 +1,17 @@
 # Release History
 
-## 5.2.0-beta.1 (2023-07-28)
+## 6.0.0 (2023-08-29)
+### Breaking Changes
+
+- Enum `FileFormat` has been removed
+- Field `FileFormat` of struct `GrantAccessData` has been removed
 
 ### Features Added
-- Support for test fakes and OpenTelemetry trace spans.
+
+- New function `*VirtualMachineScaleSetVMsClient.BeginAttachDetachDataDisks(context.Context, string, string, string, AttachDetachDataDisksRequest, *VirtualMachineScaleSetVMsClientBeginAttachDetachDataDisksOptions) (*runtime.Poller[VirtualMachineScaleSetVMsClientAttachDetachDataDisksResponse], error)`
+- New struct `AttachDataDisk`
+- New struct `AttachDetachDataDisksRequest`
+- New struct `DetachDataDisk`
 
 
 ## 5.1.0 (2023-07-28)